% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCellRegEst.R
\name{buildModel}
\alias{buildModel}
\title{Perform PLS regression}
\usage{
buildModel(
  data_4_pls,
  cv_opt = c("CV", "LOOCV")[1],
  cv_num = 10,
  component_num = NULL
)
}
\arguments{
\item{data_4_pls}{A data frame.}

\item{cv_opt}{A string.}

\item{cv_num}{A number.}

\item{component_num}{A number.}
}
\description{
Perform PLS regression
}
