% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCellReg.R
\name{cellCellReg}
\alias{cellCellReg}
\title{Perform CCPLS}
\usage{
cellCellReg(exp_mat, coord_mat, annot_mat, output_dir, HVG_extract_num = 2000)
}
\arguments{
\item{exp_mat}{A matrix object containing expression values. The rows represent cells. The columns represent genes. Row names and column names correspond to cell IDs and gene symbols, respectively.}

\item{coord_mat}{A matrix object containing cell IDs, x and y coordinates. The rows represent cells. The first column contains cell IDs and the second and subsequent columns contain x and y coordinates.}

\item{annot_mat}{A matrix object containing cell IDs and cell type labels. The rows represent cells. The first column contains cell IDs and the second column contains cell type labels.}

\item{output_dir}{A string. Full path of directory for saving outputs generated by CCPLS.}

\item{HVG_extract_num}{A number. Option for specifying extraction number of HVGs. The value is set as 2,000 in the default setting.}
}
\description{
Arguments for cellCellReg
}
