% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCellRegEst.R
\name{cellCellRegEst}
\alias{cellCellRegEst}
\title{Estimate PLS regression model}
\usage{
cellCellRegEst(
  fet_mat_orig,
  annot_score_mat,
  exp_mat_sep_list,
  fet_mat_sep_list,
  estimate_cell_type_opt = "all",
  cell_type_list,
  perm_opt = FALSE,
  component_num_list = NULL,
  dev_opt = "pls"
)
}
\arguments{
\item{fet_mat_orig}{A data frame.}

\item{annot_score_mat}{A data frame.}

\item{exp_mat_sep_list}{A list.}

\item{fet_mat_sep_list}{A list.}

\item{estimate_cell_type_opt}{A string.}

\item{cell_type_list}{A string.}

\item{perm_opt}{A logical.}

\item{component_num_list}{tentative.}

\item{dev_opt}{A string.}
}
\description{
Estimate PLS regression model
}
