% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCellRegGraph.R
\name{cellCellRegGraph}
\alias{cellCellRegGraph}
\title{Report bipartite graph}
\usage{
cellCellRegGraph(res.sel.var = res.sel.var, output_dir = output_dir)
}
\arguments{
\item{res.sel.var}{A list.}

\item{output_dir}{A string.}
}
\description{
Report bipartite graph
}
