% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCellRegHeatmap.R
\name{cellCellRegHeatmap}
\alias{cellCellRegHeatmap}
\title{Report heat map}
\usage{
cellCellRegHeatmap(res.sel.var = res.sel.var, output_dir = output_dir)
}
\arguments{
\item{res.sel.var}{A list.}

\item{output_dir}{A string.}
}
\description{
Report heat map
}
