% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCellRegMakeFet.R
\name{cellCellRegMakeFet}
\alias{cellCellRegMakeFet}
\title{Calculate neighbor cell type score}
\usage{
cellCellRegMakeFet(coord_mat = coord_mat, annot_mat = annot_mat)
}
\arguments{
\item{coord_mat}{Coordinate matrix}

\item{annot_mat}{Annotation score matrix}
}
\description{
Calculate neighbor cell type score
}
