% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCellRegSelVar.R
\name{cellCellRegSelVar}
\alias{cellCellRegSelVar}
\title{Filter coefficient}
\usage{
cellCellRegSelVar(
  res.estimate = res.estimate,
  res.sep.mat = res.sep.mat,
  HVG_extract_num = HVG_extract_num,
  dev_opt = "kmeans"
)
}
\arguments{
\item{res.estimate}{A list.}

\item{res.sep.mat}{A list.}

\item{HVG_extract_num}{A number.}

\item{dev_opt}{A string. Default is "kmeans".}
}
\description{
Filter coefficient
}
