% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCellRegSepMat.R
\name{cellCellRegSepMat}
\alias{cellCellRegSepMat}
\title{Perform preprocess}
\usage{
cellCellRegSepMat(
  exp_mat = exp_mat,
  fet_mat = fet_mat,
  annot_mat = annot_mat,
  annot_score_mat = annot_score_mat,
  HVG_opt = HVG_opt,
  HVG_extract_num = HVG_extract_num
)
}
\arguments{
\item{exp_mat}{A data frame.}

\item{fet_mat}{A data frame.}

\item{annot_mat}{A data frame.}

\item{annot_score_mat}{A data frame.}

\item{HVG_opt}{A logical.}

\item{HVG_extract_num}{A number.}
}
\description{
Perform preprocess
}
