% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCellRegSepMat.R
\name{performSeurat}
\alias{performSeurat}
\title{Perform Seurat}
\usage{
performSeurat(exp_mat, HVG_extract_num)
}
\arguments{
\item{exp_mat}{A data frame.}

\item{HVG_extract_num}{A number.}
}
\description{
Perform Seurat
}
