% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCellRegSelVar.R
\name{returnDelFetInd}
\alias{returnDelFetInd}
\title{Get index of feature for deleting}
\usage{
returnDelFetInd(fet_mat_norm, x_score_mat, calc_comp_num, p_value_thresh)
}
\arguments{
\item{fet_mat_norm}{A data frame.}

\item{x_score_mat}{A data frame.}

\item{calc_comp_num}{A number.}

\item{p_value_thresh}{A number.}
}
\description{
Get index of feature for deleting
}
