% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCellRegSelVar.R
\name{sigCoef}
\alias{sigCoef}
\title{Filter coefficient}
\usage{
sigCoef(res.pls, exp_mat_norm, fet_mat_norm, p_value_thresh = 0.05)
}
\arguments{
\item{res.pls}{A list.}

\item{exp_mat_norm}{A data frame.}

\item{fet_mat_norm}{A data frame.}

\item{p_value_thresh}{A number.}
}
\description{
Filter coefficient
}
