% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCellRegSelVar.R
\name{returnDelGeneInd}
\alias{returnDelGeneInd}
\title{Get index of gene for deleting}
\usage{
returnDelGeneInd(exp_mat_norm, y_score_mat, calc_comp_num, p_value_thresh)
}
\arguments{
\item{exp_mat_norm}{A data frame.}

\item{y_score_mat}{A data frame.}

\item{calc_comp_num}{A number.}

\item{p_value_thresh}{A number.}
}
\description{
Get index of gene for deleting
}
